Move eth spent functions out of the more generic HistoricalBalance
